feat: Move auth under api
